+++ components/Chat/Message/Content/Image.tsx
@@ -1,17 +1,40 @@
-import { FC, useEffect, CSSProperties } from 'react'
+import { FC, useEffect, CSSProperties, useCallback, useState } from 'react'
 import { ImageType } from '@botui/types'
 import { useMessageContext } from '../../../../hooks/use-message-context'
 import { AmplifyS3Image } from '@aws-amplify/ui-react'
+import Loading from './Loading'
+import { css } from '@emotion/react'
+
+const style = {
+  hidden: css({
+    display: 'none'
+  })
+}
 
 const imageStyle = ({ '--width': '100%' } as unknown) as CSSProperties
 
 const Image: FC = () => {
   const { message, handleUpdate } = useMessageContext()
+  const [loading, setLoading] = useState(true)
   const props = message.content.props as ImageType
   useEffect(() => {
-    handleUpdate && handleUpdate({ ...message, completed: true })
-  }, [])
+    loading && handleUpdate && handleUpdate({ ...message, completed: true })
+  }, [loading])
+  const handleLoad = useCallback(() => {
+    setLoading(false)
+  }, [setLoading])
 
-  return <AmplifyS3Image {...props} style={imageStyle} />
+  return (
+    <>
+      {loading && <Loading />}
+      <AmplifyS3Image
+        {...props}
+        style={imageStyle}
+        css={loading ? style.hidden : null}
+        handleOnLoad={handleLoad}
+        handleOnError={handleLoad}
+      />
+    </>
+  )
 }
 export default Image
