+++ package.json
@@ -25,6 +25,7 @@
     "@types/react": "^16.9.56",
     "@types/react-dom": "^16.9.9",
     "@types/react-scroll": "^1.8.2",
+    "@types/uuid": "^8.3.0",
     "@types/yup": "^0.29.9",
     "ace-builds": "^1.4.12",
     "aws-amplify": "^3.3.8",
@@ -51,6 +52,7 @@
     "typescript": "^4.0.5",
     "use-cors-state": "^1.1.1",
     "use-postal-jp": "^1.1.4",
+    "uuid": "^8.3.1",
     "yup": "^0.29.3"
   },
   "devDependencies": {
